% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{Centering datasets}
\usage{
center(dataset, desired)
}
\arguments{
\item{dataset}{The numeric vector to be centered}

\item{desired}{The numeric midpoint value around which the data will be centered}
}
\value{
A new vector containing the original data centered around the desired values
}
\description{
Some more description
}
\details{
And even more
}
