% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Rescaling datasets}
\usage{
rescale(v)
}
\arguments{
\item{v}{A numeric vector}
}
\value{
A new numeric vector, rescaled from the input vector to the range of 0 to 1
}
\description{
Rescaling datasets
}
\examples{
rescale(c(1,2,3)) # [1] 0.0 0.5 1.0
rescale(c(1,2,3,4,5)) # [1] 0.00 0.25 0.50 0.75 1.00
}
